rough type addition; will break many things [skip ci]
