+++ packages/common/types/AutoTestTypes.ts
@@ -1,25 +1,4 @@
-export interface CommitTarget {
-    /**
-     * The delivId the commit should be executed against. If the course does not
-     * have a default id and one is not specified (e.g., in a commit comment)
-     * we cannot create a CommitTarget.
-     */
-    delivId: string; // TODO: remove this from here? it's already in IContainerInput
-    repoId: string;
-
-    botMentioned: boolean; // true if explicitly mentioned
-    personId: string | null; // string is Person.id if explicily invoked, null otherwise
-
-    cloneURL: string;
-
-    commitSHA: string;
-    commitURL: string;
-
-    postbackURL: string; // where to send postback results
-    timestamp: number; // timestamp of push event
-
-    flags?: string[]; // other flags for the commit (e.g., #slient, #force)
-}
+import {ContainerInput, GradeReport} from "./ContainerTypes";
 
 /**
  * Pertinent properties from GitHub push webhook events.
@@ -44,12 +23,24 @@ export interface IFeedbackGiven {
 }
 
 /**
- * This is the result of an AutoTest run.
+ * Description of attachments that are saved in files on disk. This
+ * helps minimize database size making it easier to backup and much
+ * quicker to search and traverse (especially over the network).
+ */
+export interface IAttachment {
+    name: string; // file identifier attachment (e.g., stdio.txt)
+    path: string; // path to file (including name)
+    content_type: string;
+}
+
+/**
+ * This is the result of an AutoTest run. It is constructed by the Grader
+ * and sent back to Classy for querying on the backend as needed.
  *
  * There is some duplication in the record to enable easier querying.
  *
  */
-export interface IAutoTestResult {
+export interface AutoTestResult {
 
     /**
      * Foreign key into Deliverables.
@@ -70,79 +61,18 @@ export interface IAutoTestResult {
     commitURL: string;
     commitSHA: string; // can be used to index into the AutoTest collections (pushes, comments, & feedback)
 
-    input: IContainerInput; // Prepared by AutoTest service
-    output: IContainerOutput; // Returned by the Grader service
-}
-
-/**
- * Primary data structure that the course container is invoked with.
- */
-export interface IContainerInput {
-    delivId: string; // Specifies what delivId the Grader should execute against.
-    // TODO: rename to target: ICommitTarget
-    pushInfo: CommitTarget; // Details about the push event that led to this request.
-    containerConfig: AutoTestConfig; // Container configuration details.
-}
-
-/**
- * Primary data structure that the course container returns.
- */
-export interface IGradeReport {
-    scoreOverall: number; // must be set
-    scoreTest: number | null; // null means not valid for this report
-    scoreCover: number | null; // null means not valid for this report
-
-    // The semantics of these four categories are up to the container
-    // we only differentiate them so the report UI can render them uniquely.
-    // Set to [] for any unused property.
-    passNames: string[];
-    failNames: string[];
-    errorNames: string[];
-    skipNames: string[];
-
-    // This is the text of the feedback (in markdown) that the container wants
-    // to return to the user.
-    feedback: string;
-
-    // Enables custom values to be returned to the UI layer.
-    // PLEASE: do not store large objects in here or it will
-    // significantly impact the performance of the dashboard.
-    // Use attachments instead for large bits of data you wish
-    // to persist.
-    custom: {};
+    input: ContainerInput; // Prepared by AutoTest service
+    output: ContainerOutput; // Returned by the Grader service
 }
 
 /**
  * Primary data structure that is returned by a Grader.
  */
-export interface IContainerOutput {
+export interface ContainerOutput {
     timestamp: number; // time when complete
-    report: IGradeReport;
+    report: GradeReport;
     postbackOnComplete: boolean;
     attachments: IAttachment[];
-    state: string; // enum: SUCCESS, FAIL, TIMEOUT, INVALID_REPORT
+    state: string; // SUCCESS, FAIL, TIMEOUT, INVALID_REPORT, LINT, COMPILE // TODO: move to GradeReport
     custom: {};
-}
-
-/**
- * Description of attachments that are saved in files on disk. This
- * helps minimize database size making it easier to backup and much
- * quicker to search and traverse (especially over the network).
- */
-export interface IAttachment {
-    name: string; // file identifier attachment (e.g., stdio.txt)
-    path: string; // path to file (including name)
-    content_type: string;
-}
-
-/**
- * Description of the configuration parameters for the AutoTest container.
- * These can be specified per-deliverable in the Portal UI.
- */
-export interface AutoTestConfig {
-    dockerImage: string; // docker image that should be launchd
-    studentDelay: number; // minimum interval between student requests in seconds
-    maxExecTime: number; // maximum execution time for the container before it will be timed out in seconds
-    regressionDelivIds: string[]; // other deliverables that should be run alongside any request
-    custom: {}; // custom parameters
 }
